payment_api_v1/requirements.txt
Django==2.1.7
django-money==0.14.4
django-model-utils==3.1.2
psycopg2-binary==2.7.7
djangorestframework==3.9.1
https://github.com/celery/celery/zipball/master#egg=celery
redis==3.2.0